+++ FitChain/App.tsx
@@ -1,64 +1,47 @@
-/**
- * FitChain - Step Counter App
- * A React Native app to count steps with proper permissions
- *
- * @format
- */
+import '@walletconnect/react-native-compat';
 
-import React, { useState, useCallback, useMemo } from 'react';
-import { SafeAreaView, StatusBar, StyleSheet, useColorScheme } from 'react-native';
-import { NavigationContainer } from '@react-navigation/native';
-import { COLORS } from './src/constants/theme';
-import TabNavigator from './src/componenets/navigation/TabNavigator';
-import SplashScreen from './src/screens/SplashScreen';
-import { WagmiProvider } from 'wagmi';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { wagmiConfig } from './src/config/wagmi';
-import { AppKit } from '@reown/appkit-wagmi-react-native';
+// Add BackHandler polyfill
+import { ScrollView, StyleSheet, Text, View } from 'react-native';
 
+import {WagmiProvider} from 'wagmi';
+import {QueryClient, QueryClientProvider} from '@tanstack/react-query';
+import {AppKit, AppKitButton} from '@reown/appkit-wagmi-react-native';
+import {wagmiConfig} from './src/config/wagmi';
+import {Account} from './src/componenets/Account';
+import {FitnessRewards} from './src/componenets/FitnessRewards';
 const queryClient = new QueryClient();
 
-
-function App(): React.JSX.Element {
-  const isDarkMode = useColorScheme() === 'dark';
-  const [showSplash, setShowSplash] = useState(true);
-
-  const backgroundStyle = useMemo(() => ({
-    backgroundColor: isDarkMode ? COLORS.BACKGROUND_DARK : COLORS.BACKGROUND_LIGHT,
-    flex: 1,
-  }), [isDarkMode]);
-
-  const handleSplashFinish = useCallback(() => {
-    setShowSplash(false);
-  }, []);
-
-  if (showSplash) {
-    return <SplashScreen onFinish={handleSplashFinish} />;
-  }
-
+export default function App() {
   return (
     <WagmiProvider config={wagmiConfig}>
-    <QueryClientProvider client={queryClient}>
-    <SafeAreaView style={[backgroundStyle, styles.container]}>
-      <StatusBar
-        barStyle={isDarkMode ? 'light-content' : 'dark-content'}
-        backgroundColor={backgroundStyle.backgroundColor}
-      />
-      <NavigationContainer>
-        <TabNavigator />
-      </NavigationContainer>
-      <AppKit />
-
-    </SafeAreaView>
-    </QueryClientProvider>
+      <QueryClientProvider client={queryClient}>
+        <View style={styles.container}>
+          <ScrollView contentContainerStyle={styles.scrollContent}>
+            <Text style={styles.header}>FitChain</Text>
+            <AppKitButton />
+            <Account />
+            <FitnessRewards />
+          </ScrollView>
+        </View>
+        <AppKit />
+      </QueryClientProvider>
     </WagmiProvider>
   );
 }
 
 const styles = StyleSheet.create({
   container: {
-    position: 'relative',
+    flex: 1,
+    backgroundColor: '#f5f5f5',
   },
+  scrollContent: {
+    paddingBottom: 30,
+  },
+  header: {
+    fontSize: 32,
+    fontWeight: 'bold',
+    textAlign: 'center',
+    marginVertical: 20,
+    color: '#333',
+  }
 });
-
-export default App;
