+++ FitChain/package.json
@@ -12,22 +12,15 @@
   "dependencies": {
     "@react-native-async-storage/async-storage": "^2.1.2",
     "@react-native-community/netinfo": "^11.4.1",
-    "@react-navigation/bottom-tabs": "^7.2.1",
-    "@react-navigation/native": "^7.0.15",
     "@reown/appkit-wagmi-react-native": "^1.2.2",
     "@tanstack/react-query": "^5.67.1",
     "@walletconnect/react-native-compat": "^2.19.0",
     "react": "19.0.0",
     "react-native": "0.78.0",
-    "react-native-gesture-handler": "^2.24.0",
     "react-native-get-random-values": "^1.11.0",
     "react-native-modal": "^13.0.1",
-    "react-native-reanimated": "^3.17.1",
-    "react-native-safe-area-context": "^5.3.0",
-    "react-native-screens": "^4.9.1",
     "react-native-sensors": "^7.3.6",
     "react-native-svg": "^15.11.2",
-    "react-native-vector-icons": "^10.2.0",
     "viem": "^2.23.7",
     "wagmi": "^2.14.12"
   },
