--- FitChain/src/config/app.config.ts
@@ -1,15 +1,0 @@
-export const APP_CONFIG = {
-  VERSION: '1.0.0',
-  BUILD_NUMBER: '1',
-  API_URL: 'https://api.fitchain.com', // Replace with actual API URL
-  STORAGE_KEYS: {
-    USER_SETTINGS: '@fitchain_user_settings',
-    STEP_DATA: '@fitchain_step_data',
-  },
-  DEFAULT_SETTINGS: {
-    dailyGoal: 10000,
-    notifications: true,
-    darkMode: false,
-    units: 'metric',
-  },
-};
