--- FitChain/src/screens/AboutScreen.tsx
@@ -1,183 +1,0 @@
-import React from 'react';
-import {
-  ScrollView,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-  Linking,
-} from 'react-native';
-
-const AboutScreen: React.FC = () => {
-  return (
-    <ScrollView style={styles.container}>
-      <View style={styles.header}>
-        <Text style={styles.appName}>FitChain</Text>
-        <Text style={styles.version}>Version 1.0.0</Text>
-      </View>
-
-      <View style={styles.section}>
-        <Text style={styles.sectionTitle}>About the App</Text>
-        <Text style={styles.description}>
-          FitChain is a step counter application designed to help you track your daily physical activity.
-          The app uses your device's built-in sensors to count steps, calculate distance traveled, and
-          estimate calories burned.
-        </Text>
-      </View>
-
-      <View style={styles.section}>
-        <Text style={styles.sectionTitle}>Features</Text>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Step counting with device sensors</Text>
-        </View>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Distance and calorie tracking</Text>
-        </View>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Background tracking capability</Text>
-        </View>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Milestone notifications</Text>
-        </View>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Customizable settings</Text>
-        </View>
-      </View>
-
-      <View style={styles.section}>
-        <Text style={styles.sectionTitle}>How It Works</Text>
-        <Text style={styles.description}>
-          FitChain uses your device's accelerometer to detect the characteristic motion patterns of walking.
-          When the app detects a step pattern, it increments your step count and calculates related metrics
-          like distance and calories.
-        </Text>
-        <Text style={[styles.description, { marginTop: 10 }]}>
-          For the most accurate results, keep your phone in your pocket or bag while walking.
-        </Text>
-      </View>
-
-      <View style={styles.section}>
-        <Text style={styles.sectionTitle}>Permissions</Text>
-        <Text style={styles.description}>
-          FitChain requires the following permissions to function properly:
-        </Text>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Activity Recognition - To count your steps</Text>
-        </View>
-        <View style={styles.featureItem}>
-          <Text style={styles.featureBullet}>•</Text>
-          <Text style={styles.featureText}>Location (optional) - For more accurate distance calculation</Text>
-        </View>
-      </View>
-
-      <View style={styles.section}>
-        <Text style={styles.sectionTitle}>Contact Us</Text>
-        <TouchableOpacity 
-          style={styles.contactButton}
-          onPress={() => Linking.openURL('mailto:support@fitchain.com')}
-        >
-          <Text style={styles.contactButtonText}>Email Support</Text>
-        </TouchableOpacity>
-      </View>
-
-      <View style={styles.footer}>
-        <Text style={styles.footerText}>© 2025 FitChain</Text>
-        <Text style={styles.footerText}>All Rights Reserved</Text>
-      </View>
-    </ScrollView>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f5f5f5',
-    padding: 16,
-  },
-  header: {
-    alignItems: 'center',
-    marginVertical: 24,
-  },
-  appName: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    color: '#4CAF50',
-    marginBottom: 8,
-  },
-  version: {
-    fontSize: 16,
-    color: '#666',
-  },
-  section: {
-    backgroundColor: 'white',
-    borderRadius: 12,
-    padding: 16,
-    marginBottom: 20,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0,
-      height: 2,
-    },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  sectionTitle: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    color: '#333',
-    marginBottom: 12,
-  },
-  description: {
-    fontSize: 16,
-    color: '#555',
-    lineHeight: 24,
-  },
-  featureItem: {
-    flexDirection: 'row',
-    marginBottom: 8,
-    alignItems: 'flex-start',
-  },
-  featureBullet: {
-    fontSize: 18,
-    color: '#4CAF50',
-    marginRight: 8,
-    lineHeight: 24,
-  },
-  featureText: {
-    fontSize: 16,
-    color: '#555',
-    flex: 1,
-    lineHeight: 24,
-  },
-  contactButton: {
-    backgroundColor: '#4CAF50',
-    borderRadius: 8,
-    padding: 12,
-    alignItems: 'center',
-    marginTop: 8,
-  },
-  contactButtonText: {
-    color: 'white',
-    fontWeight: 'bold',
-    fontSize: 16,
-  },
-  footer: {
-    marginTop: 8,
-    marginBottom: 40,
-    alignItems: 'center',
-  },
-  footerText: {
-    color: '#666',
-    fontSize: 14,
-    marginBottom: 4,
-  },
-});
-
-export default AboutScreen;
