--- FitChain/src/screens/SplashScreen.tsx
@@ -1,96 +1,0 @@
-import React, { useEffect } from 'react';
-import {
-  View,
-  Text,
-  StyleSheet,
-  Animated,
-  Easing,
-} from 'react-native';
-
-interface SplashScreenProps {
-  onFinish: () => void;
-}
-
-const SplashScreen: React.FC<SplashScreenProps> = ({ onFinish }) => {
-  const scaleValue = new Animated.Value(0);
-  const opacityValue = new Animated.Value(0);
-
-  useEffect(() => {
-    // Animation sequence
-    Animated.sequence([
-      // Fade in and scale up
-      Animated.parallel([
-        Animated.timing(scaleValue, {
-          toValue: 1,
-          duration: 800,
-          easing: Easing.out(Easing.cubic),
-          useNativeDriver: true,
-        }),
-        Animated.timing(opacityValue, {
-          toValue: 1,
-          duration: 800,
-          useNativeDriver: true,
-        }),
-      ]),
-      // Wait for a moment
-      Animated.delay(1000),
-      // Fade out
-      Animated.timing(opacityValue, {
-        toValue: 0,
-        duration: 500,
-        useNativeDriver: true,
-      }),
-    ]).start(() => {
-      // Call onFinish when animation is complete
-      onFinish();
-    });
-  }, [onFinish, opacityValue, scaleValue]);
-
-  return (
-    <View style={styles.container}>
-      <Animated.View
-        style={[
-          styles.logoContainer,
-          {
-            opacity: opacityValue,
-            transform: [{ scale: scaleValue }],
-          },
-        ]}
-      >
-        <Text style={styles.logo}>FC</Text>
-        <Text style={styles.appName}>FitChain</Text>
-        <Text style={styles.tagline}>Track your steps, stay active</Text>
-      </Animated.View>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#4CAF50',
-  },
-  logoContainer: {
-    alignItems: 'center',
-  },
-  logo: {
-    fontSize: 72,
-    fontWeight: 'bold',
-    color: 'white',
-    marginBottom: 10,
-  },
-  appName: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    color: 'white',
-    marginBottom: 10,
-  },
-  tagline: {
-    fontSize: 16,
-    color: 'rgba(255, 255, 255, 0.8)',
-  },
-});
-
-export default SplashScreen;
