+++ FitChain/tsconfig.json
@@ -1,6 +1,3 @@
 {
-  "compilerOptions": {
-    "jsx": "react"
-  },
   "extends": "@react-native/typescript-config/tsconfig.json"
 }
